added non-threaded option for processing, since the threaded version wasn't saving the feature maps properly

+++ src/processing/grotoap_to_tfrecords.py
@@ -17,30 +17,17 @@ tf.app.flags.DEFINE_string('grotoap_dir', '', 'top level directory containing gr
 tf.app.flags.DEFINE_string('out_dir', '', 'export tf protos')
 tf.app.flags.DEFINE_string('load_vocab', '', 'directory containing embedding vocab files to load')
 tf.app.flags.DEFINE_boolean('use_lexicons', False, 'use string lexicon features')
-tf.app.flags.DEFINE_integer('num_threads', 12, 'max number of threads to use for parallel processing')
+tf.app.flags.DEFINE_integer('num_threads', 1, 'max number of threads to use for parallel processing')
 # tf.app.flags.DEFINE_boolean('padding', 0, '0: no padding, 1: 0 pad to the right of seq, 2: 0 pad to the left')
 # tf.app.flags.DEFINE_boolean('normalize_digits', False, 'map all digits to 0')
 # tf.app.flags.DEFINE_boolean('lowercase', False, 'whether to lowercase')
 tf.app.flags.DEFINE_boolean('debug', False, 'print debugging output')
+tf.app.flags.DEFINE_boolean('bilou', False, 'encode the word labels in BILOU format')
 
 FLAGS = tf.app.flags.FLAGS
 
 # PAD_STR = "<PAD>"
 OOV_STR = "<OOV>"
-
-# Helpers for creating Example objects (as defined in Pat's code)
-# feature = tf.train.Feature
-# sequence_example = tf.train.SequenceExample
-#
-#
-# def features(d): return tf.train.Features(feature=d)
-# def int64_feature(v): return feature(int64_list=tf.train.Int64List(value=v))
-# def feature_list(l): return tf.train.FeatureList(feature=l)
-# def feature_lists(d): return tf.train.FeatureLists(feature_list=d)
-
-# Emma's helpers
-# def _int64_feature(value): return tf.train.Feature(int64_list=tf.train.Int64List(value=value))
-# def _float64_feature(value): return tf.train.Feature(float64_list=tf.train.Float64List(value=value))
 
 
 # TODO: do we need these? i think pat only used them to keep track of the processing progress
@@ -187,9 +174,7 @@ def make_example(writer, page, update_vocab, update_chars):
     page_lens = [len(word_tups)]
     tok_lens = []
 
-    last_label = "O"
     labels = []
-    current_tag = ''
     char_start = 0
 
     # iterate over the words in the sequence and extract features, update vocab maps
@@ -211,12 +196,10 @@ def make_example(writer, page, update_vocab, update_chars):
             if FLAGS.debug:
                 print(token_normalized + " not in vocab.")
             if update_vocab:
-                print("adding to vocab")
+                # print("adding to vocab")
                 token_map[token_normalized] = len(token_map)
                 token_int_str_map[token_map[token_normalized]] = token_normalized
 
-        # TODO: generate BIO labels
-        # label_bilou = generate_bio(label, last_label)
         label_bilou = label
 
         # get bounding box info (and other geometrical features?)
@@ -270,10 +253,10 @@ def make_example(writer, page, update_vocab, update_chars):
     y_bins = np.linspace(0, y_coords.max(), num=4)
     y_coords = np.digitize(y_coords, y_bins)
 
-    # TODO: why are we intmapping the labels here? is it because of earlier BIO processing?
-
+    # move this intmapping elsewhere?
     for label in labels:
         if label not in label_map:
+            print("label %s not in vocab. adding..." % label)
             label_map[label] = len(label_map)
             label_int_str_map[label_map[label]] = label
 
@@ -312,6 +295,11 @@ def make_example(writer, page, update_vocab, update_chars):
                       lines,
                       zones)
 
+    print(len(label_map))
+    print(len(token_map))
+    print(len(char_map))
+    print(len(shape_map))
+
     return len(word_tups), oov_count, 1
 
 
@@ -321,19 +309,6 @@ def doc_to_examples(in_file, writer):
     update_vocab = True
     update_chars = True
 
-    # # add the padding string to the token map
-    # token_map[PAD_STR] = len(token_map)
-    # token_int_str_map[token_map[PAD_STR]] = PAD_STR
-    # # add the padding string to the character map
-    # char_map[PAD_STR] = len(char_map)
-    # char_int_str_map[char_map[PAD_STR]] = PAD_STR
-    # # add the padding string to the shape map
-    # shape_map[PAD_STR] = len(shape_map)
-    # # if we want to predict the padding string, add it to the label map as well
-    # if FLAGS.predict_pad:
-    #     label_map[PAD_STR] = len(label_map)
-    #     label_int_str_map[label_map[PAD_STR]] = PAD_STR
-
     # add out of vocab string to the token, character maps
     token_map[OOV_STR] = len(token_map)
     token_int_str_map[token_map[OOV_STR]] = OOV_STR
@@ -342,13 +317,10 @@ def doc_to_examples(in_file, writer):
     char_int_str_map[char_map[OOV_STR]] = OOV_STR
 
     try:
-        # in_f, out_path = in_out
-        # writer = tf.python_io.TFRecordWriter(out_path)
-        # print('Converting %s to %s' % (in_f, out_path))
         print('Converting %s ' % in_file)
         doc = parse_doc(in_file)
-        # for page in doc.pages:
-        #     make_example(writer, page, label_map, token_map, shape_map, char_map, update_vocab, update_chars)
+        if FLAGS.bilou:
+            words_to_bilou(doc)
         # just start by trying first page only
         num_words, oov_count, _ = make_example(writer, doc.pages[0], update_vocab, update_chars)
         # writer.close()
@@ -358,20 +330,13 @@ def doc_to_examples(in_file, writer):
         return 'KeyboardException'
 
 def dir_to_examples(root_dir, dir_out):
-    # load vocab (from the embeddings file)
-    if FLAGS.load_vocab != '':
-        print("loading vocab...")
-        update_vocab = False
-        with open(FLAGS.load_vocab, 'r') as f:
-            for line in f.readlines():
-                word = line.strip().split(" ")[0]
-                # token_map[word] = int(idx)
-                # word = line.strip().split("\t")[0]
-                if word not in token_map:
-                    # print("adding word %s" % word)
-                    token_map[word] = len(token_map)
-                    token_int_str_map[token_map[word]] = word
+    '''
+    Convert a directory of TrueViz document files to one TFRecord file containing examples from all of them
 
+    :param root_dir: directory of files
+    :param dir_out: target TFRecord file
+    :return:
+    '''
     (dir_path, out_path) = dir_out
     print("Converting directory %s to TFRecord %s" % (dir_path, out_path))
     writer = tf.python_io.TFRecordWriter(out_path)
@@ -386,64 +351,57 @@ def dir_to_examples(root_dir, dir_out):
                 tot_words += num_words
                 tot_oov += oov_count
     print("Done with directory %s" % dir_path)
-    coverage = 1 - tot_oov/tot_words
+    coverage = 1 - tot_oov/len(token_int_str_map)
     print("Embeddings coverage: %f" % coverage)
     writer.close()
 
 
-def grotoap_to_examples():
+def grotoap_to_examples(use_threads=False):
+    '''
+    Convert entire GROTOAP set to TFRecords (iterate over the subdirectors 00, 01, ..., 99 and create a TFRecord for each)
+    :return:
+    '''
+    # load vocab (from the embeddings file)
+    if FLAGS.load_vocab != '':
+        print("loading vocab...")
+        update_vocab = False
+        with open(FLAGS.load_vocab, 'r') as f:
+            for line in f.readlines():
+                word = line.strip().split(" ")[0]
+                if word not in token_map:
+                    # print("adding word %s" % word)
+                    token_map[word] = len(token_map)
+                    token_int_str_map[token_map[word]] = word
+
     if not os.path.exists(FLAGS.out_dir):
         os.makedirs(FLAGS.out_dir)
 
-    # train_writer = tf.python_io.TFRecordWriter(FLAGS.out_dir + os.sep + 'grotoap.train.proto')
-    # test_writer = tf.python_io.TFRecordWriter(FLAGS.out_dir + os.sep + 'grotoap.test.proto')
-    # dev_writer = tf.python_io.TFRecordWriter(FLAGS.out_dir + os.sep + 'grotoap.dev.proto')
-
-    # in_files = []
     in_dirs = []
-    out_files = []
-    # TODO: point this at the grotoap directories, have one out file for each directory?
+
+    # should point this at the directory containing the numbered subdirectories
     for root, subdirs, files in os.walk(FLAGS.grotoap_dir):
         for subdir in subdirs:
             in_dirs.append(subdir)
-        # for file in files:
-        #     if '.cxml' in file:
-        #         # filepath = FLAGS.grotoap_dir + os.sep + subdir + file
-        #         filepath = subdir + os.sep + file
-        #         in_files.append(filepath)
 
-    # print(in_files)
-    # write examples to 10 TFRecord files
-    # for i,f in enumerate(in_files):
-    #     # TODO: this doesn't work, it'll just overwrite the files every time a new writer is created...
-    #     # out_files.append('%s/%s.proto' % (FLAGS.out_dir, i % 10))
-    #     if i < 10000:
-    #         out_files.append(FLAGS.out_dir + os.sep + 'grotoap.train.proto')
-    #     elif i < 11500:
-    #         out_files.append(FLAGS.out_dir + os.sep + 'grotoap.test.proto')
-    #     else:
-    #         out_files.append(FLAGS.out_dir + os.sep + 'grotoap.dev.proto')
-
-    # print(out_files)
-
-    # in_files = sorted(glob.glob(FLAGS.trueviz_in_files))
     out_files = ['%s/%s.proto' % (FLAGS.out_dir, in_f.split('/')[-1]) for in_f in in_dirs]
 
     total_dirs = len(in_dirs)
 
-    print('Starting file process threads using %d threads' % FLAGS.num_threads)
-    pool = multiprocessing.Pool(FLAGS.num_threads)
-    try:
-        # pool.map_async(partial(doc_to_examples, total_docs), zip(in_files, out_files)).get(999999)
-        pool.map_async(partial(dir_to_examples, FLAGS.grotoap_dir), zip(in_dirs, out_files)).get(999999)
-        pool.close()
-        pool.join()
-    except KeyboardInterrupt:
-        pool.terminate()
-
-    # train_writer.close()
-    # test_writer.close()
-    # dev_writer.close()
+    # TODO threading doesn't save the maps correctly
+    # TODO I think if we want to use multithreading we need to iterate over the data once to create the maps
+    if use_threads:
+        print('Starting file process threads using %d threads' % FLAGS.num_threads)
+        pool = multiprocessing.Pool(FLAGS.num_threads)
+        try:
+            # pool.map_async(partial(doc_to_examples, total_docs), zip(in_files, out_files)).get(999999)
+            pool.map_async(partial(dir_to_examples, FLAGS.grotoap_dir), zip(in_dirs, out_files)).get(999999)
+            pool.close()
+            pool.join()
+        except KeyboardInterrupt:
+            pool.terminate()
+    else:
+        for (in_dir, out_file) in zip(in_dirs, out_files):
+            dir_to_examples(FLAGS.grotoap_dir, (in_dir, out_file))
 
 def export_maps():
     # export the string->int maps to file
@@ -454,8 +412,8 @@ def export_maps():
 
     print('exporting string->int maps')
     for f_str, id_map in [('label', label_map), ('token', token_map), ('shape', shape_map), ('char', char_map)]:
+        print(f_str)
         with codecs.open(FLAGS.out_dir + '/' + f_str + '.txt', 'w', 'utf-8') as f:
-        # with open(FLAGS.out_dir + '/' + f_str + '.txt', 'w') as f:
             [f.write(s + '\t' + str(i) + '\n') for (s, i) in id_map.items()]
 
 def main(argv):
@@ -474,4 +432,4 @@ def main(argv):
 if __name__ == '__main__':
     tf.app.run()
 
-# python grotoap_to_tfrecords.py --out_dir $DATA_DIR --load_vocab /iesl/canvas/mmcmahon/embeddings/PubMed-w2v.txt
+# python grotoap_to_tfrecords.py --out_dir $DATA_DIR --load_vocab /iesl/canvas/mmcmahon/embeddings/PubMed-w2v.txt --bilou
